requirements.txt
pyzbar 
psycopg2-binary  
Pillow 
pytz 
streamlit 
plotly 
pydeck
pywin32